+++ src/Sequence.js
@@ -889,15 +889,19 @@ class Sequence extends CGObject {
       ctx.fillStyle = this.color.rgbaString;
       ctx.font = this.font.cssScaled(scaleFactor);
       ctx.textAlign = 'center';
-      ctx.textBaseline = 'middle';
+      // ctx.textBaseline = 'middle';
+      ctx.textBaseline = 'alphabetic'; // The default baseline works best across canvas and svg
+      const yOffset = (this.font.height * scaleFactor / 2) - 1;
       // Distance from the center of the backbone to place sequence text
       const centerOffsetDiff = ((this.bpSpacing / 2) + this.bpMargin) * scaleFactor;
       for (let i = 0, len = range.length; i < len; i++) {
         let origin = this.canvas.pointForBp(bp, centerOffset + centerOffsetDiff);
         // if (i == 0) { console.log(bp, origin)}
-        ctx.fillText(seq[i], origin.x, origin.y);
+        // ctx.fillText(seq[i], origin.x, origin.y);
+        ctx.fillText(seq[i], origin.x, origin.y + yOffset);
         origin = this.canvas.pointForBp(bp, centerOffset - centerOffsetDiff);
-        ctx.fillText(complement[i], origin.x, origin.y);
+        // ctx.fillText(complement[i], origin.x, origin.y);
+        ctx.fillText(complement[i], origin.x, origin.y + yOffset);
         bp++;
       }
       ctx.restore();